remove rouge space

+++ spec_integration/spec_helper.rb
@@ -10,12 +10,12 @@ def example_config
                                      {1..40   => 'i1-i77'},
                                      {41..46  => '1-40'},
                                      {1..40   =>  '47-86'},
-                                     {47..86  => ' 1-40 = 1. & 1. fixed one-to-one'}],
+                                     {47..86  => '1-40 = 1. & 1. fixed one-to-one'}],
                     :special      => {:linear => '47-86',
                     :weight_limit => 1.00,
                     :selected     => '1-86'}}
 end
 
 at_exit do
-  FileUtils.rm_f(Dir.glob("#{TLearn::Config::WORKING_DIR}/*"))
+  FileUtils.rm_f(Dir.glob("#{TLearn::Config::WORKING_DIR}/*")) unless ENV['KEEP_DATA']
 end
